+++ src/Sulu/Bundle/ContactBundle/Entity/ContactRepository.php
@@ -392,4 +392,111 @@ class ContactRepository extends EntityRepository
             return;
         }
     }
+
+    /**
+     * Returned filtered contacts.
+     * when pagination is active the result count is pageSize + 1 to determine has next page.
+     *
+     * @param array $filters array of filters: tags, tagOperator
+     * @param int $page
+     * @param int $pageSize
+     * @param int $limit
+     *
+     * @return Contact[]
+     */
+    public function findByFilters($filters, $page, $pageSize, $limit)
+    {
+        $queryBuilder = $this->createQueryBuilder('c')
+            ->addSelect('c')
+            ->addSelect('emails')
+            ->addSelect('phones')
+            ->addSelect('faxes')
+            ->addSelect('urls')
+            ->addSelect('tags')
+            ->addSelect('categories')
+            ->addSelect('translations')
+            ->leftJoin('c.emails', 'emails')
+            ->leftJoin('c.phones', 'phones')
+            ->leftJoin('c.faxes', 'faxes')
+            ->leftJoin('c.urls', 'urls')
+            ->leftJoin('c.tags', 'tags')
+            ->leftJoin('c.categories', 'categories')
+            ->leftJoin('categories.translations', 'translations')
+            ->where('c.id IN (:ids)')
+            ->orderBy('c.id', 'ASC');
+
+        $query = $queryBuilder->getQuery();
+        $ids = array_map(
+            function ($item) {
+                return $item['id'];
+            },
+            $this->findByFiltersIds($filters, $page, $pageSize, $limit)
+        );
+
+        $query->setParameter('ids', $ids);
+
+        return $query->getResult();
+    }
+
+    /**
+     * @param array $filters array of filters: tags, tagOperator
+     * @param int $page
+     * @param int $pageSize
+     * @param int $limit
+     *
+     * @return array
+     */
+    private function findByFiltersIds($filters, $page, $pageSize, $limit)
+    {
+        $parameter = [];
+
+        $queryBuilder = $this->createQueryBuilder('c')
+            ->select('c.id');
+
+        if (isset($filters['tags']) && count($filters['tags']) > 0 && strtolower($filters['tagOperator']) === 'or') {
+            $queryBuilder->join('c.tags', 'tags')
+                ->where('tags.id IN (:tags)');
+
+            $parameter['tags'] = $filters['tags'];
+        }
+
+        if (isset($filters['tags']) && count($filters['tags']) > 0 && strtolower($filters['tagOperator']) === 'and') {
+            $expr = $queryBuilder->expr()->andX();
+
+            $len = count($filters['tags']);
+            for ($i = 0; $i < $len; ++$i) {
+                $queryBuilder->join('c.tags', 'tags' . $i);
+
+                $expr->add($queryBuilder->expr()->eq('tags' . $i . '.id', ':tag' . $i));
+
+                $parameter['tag' . $i] = $filters['tags'][$i];
+            }
+            $queryBuilder->andWhere($expr);
+        }
+
+        $query = $queryBuilder->getQuery();
+        foreach ($parameter as $name => $value) {
+            $query->setParameter($name, $value);
+        }
+
+        if ($page !== null && $pageSize > 0) {
+            $pageOffset = ($page - 1) * $pageSize;
+            $restLimit = $limit - $pageOffset;
+
+            // if limitation is smaller than the page size then use the rest limit else use page size plus 1 to
+            // determine has next page
+            $maxResults = ($limit !== null && $pageSize > $restLimit ? $restLimit : ($pageSize + 1));
+
+            if ($maxResults <= 0) {
+                return [];
+            }
+
+            $query->setMaxResults($maxResults);
+            $query->setFirstResult($pageOffset);
+        } elseif ($limit !== null) {
+            $query->setMaxResults($limit);
+        }
+
+        return $query->getScalarResult();
+    }
 }