+++ src/Sulu/Bundle/SnippetBundle/Content/SnippetContent.php
@@ -157,7 +157,7 @@ class SnippetContent extends ComplexContentType
     /**
      * {@inheritdoc}
      */
-    public function getDefaultParams()
+    public function getDefaultParams(PropertyInterface $property = null)
     {
         return [];
     }